[PP][OPS-11157] move test only pages to seperate package
