+++ test/pagespecs/pages/testonly/PtaSignInPage.scala
@@ -14,33 +14,33 @@
  * limitations under the License.
  */
 
-package pagespecs.pages
+package pagespecs.pages.testonly
 
 import org.openqa.selenium.WebDriver
 import pagespecs.pagesupport.{ContentExpectation, Page, PageUtil}
 
-class GovUkRouteInPage(baseUrl: String)(implicit webDriver: WebDriver) extends Page(
+class PtaSignInPage(baseUrl: String)(implicit webDriver: WebDriver) extends Page(
   baseUrl,
-  path = "/get-an-income-tax-refund/test-only/gov-uk-route-in"
+  path = "/get-an-income-tax-refund/test-only/pta-sign-in"
 ) {
 
-  override def expectedH1: String = "Tax overpayments and underpayments"
+  override def expectedH1: String = "Personal tax account sign in"
 
   def assertPageIsDisplayed(): Unit = withPageClue {
     PageUtil.assertPage(
-      path                = path,
-      h1                  = expectedH1,
-      title               = PageUtil.standardTitleForTestOnlyPages,
-      serviceName         = "Test Only - Claim an income tax refund",
-      contentExpectations = ContentExpectation(
+      path        = path,
+      h1          = expectedH1,
+      title       = PageUtil.standardTitleForTestOnlyPages,
+      serviceName = "Test Only - Claim an income tax refund",
+      ContentExpectation(
         atXpath       = PageUtil.Xpath.mainContent,
         expectedLines =
           """
-            |Tax overpayments and underpayments
-            |I'm the stub for Gov UK page where you can start the journey.
+            |Personal tax account sign in
+            |I'm a stub representing Personal tax account sign in page.
             |""".stripMargin
       )
     )
+    ()
   }
-
 }
